refactor(auth): improve component/file organization
