+++ src/features/auth/constants/auth-flow-types.ts
@@ -1,0 +1,4 @@
+export enum AuthFlowType {
+	EmailVerification = 'email-verification',
+	GoogleOauth = 'google-oauth',
+}
