+++ src/features/auth/routes/index.tsx
@@ -2,23 +2,26 @@ import { Route, Routes } from 'react-router-dom';
 import { Suspense } from 'react';
 import { AuthFallback } from '../components/auth-fallback';
 import { lazyImport } from '@/utils';
-const { Authentication } = lazyImport(() => import('./authentication'), 'Authentication');
+const { Login } = lazyImport(() => import('./login'), 'Login');
+const { Register } = lazyImport(() => import('./register'), 'Register');
 const { ResetPassword } = lazyImport(() => import('./reset-password'), 'ResetPassword');
 const { ResetSuccess } = lazyImport(() => import('./reset-success'), 'ResetSuccess');
 const { Welcome } = lazyImport(() => import('./welcome'), 'Welcome');
 
 export const AuthRoutes = () => {
 	return (
-		<Suspense fallback={<AuthFallback />}>
-			<Routes>
-				<Route path="register" element={<Authentication isRegisterForm={true} />} />
-				<Route path="login" element={<Authentication isRegisterForm={false} />} />
-				<Route path="reset-password/*" element={<ResetPassword />} />
-				<Route path="reset-password/success" element={<ResetSuccess />} />
-				<Route path="welcome/*" element={<Welcome />} />
-				<Route path="*" index element={<Authentication isRegisterForm={false} />} />
-			</Routes>
-		</Suspense>
+		<div style={{ backgroundColor: 'var(--color-bg-primary)', minHeight: 'var(--full-height)' }}>
+			<Suspense fallback={<AuthFallback />}>
+				<Routes>
+					<Route path="register" element={<Register />} />
+					<Route path="login" element={<Login />} />
+					<Route path="reset-password/*" element={<ResetPassword />} />
+					<Route path="reset-password/success" element={<ResetSuccess />} />
+					<Route path="welcome/*" element={<Welcome />} />
+					<Route path="*" index element={<Login />} />
+				</Routes>
+			</Suspense>
+		</div>
 	);
 };
